Move functions in tanimoto_score_calculation.py to spectrum_pair_selection.py or plotting_wrapper_functions.py
