+++ tests/test_spectrum_pair_selection.py
@@ -5,9 +5,7 @@ from scipy.sparse import coo_array
 from ms2deepscore.train_new_model.spectrum_pair_selection import (
     SelectedCompoundPairs, compute_jaccard_similarity_per_bin,
     convert_selected_pairs_per_bin_to_coo_array, fix_bias,
-    get_nr_of_pairs_needed_to_fix_bias)
-from ms2deepscore.train_new_model.tanimoto_score_calculation import \
-    select_inchi_for_unique_inchikeys
+    get_nr_of_pairs_needed_to_fix_bias, select_inchi_for_unique_inchikeys)
 
 
 @pytest.fixture
